Simulated lights and Time consideration
Included different pictures to simulate different stages of light depending on the given voice command.
Also included the consideration of time when deciding what kind of light to turn on (Making it so that if the lights are commanded to be turned on and its after 17:00 then the light would be warmer but if its before that then it would be lit brightly.